# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface User {
  id: string;
  username: string;
  /** @nullable */
  email?: string | null;
  createdAt: string;
  karma?: number;
}

export interface CurrentUserResponse {
  user: User | null;
}

export interface SignUpInput {
  /**
     * @minLength 3
     * @maxLength 24
     * @pattern ^[A-Za-z0-9_-]+$
     */
  username: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}

export interface AuthResponse {
  user: User;
}

export interface Author {
  username: string;
  id?: string;
}

export type PostListItemType = typeof PostListItemType[keyof typeof PostListItemType];


export const PostListItemType = {
  discuss: 'discuss',
  ask: 'ask',
  show: 'show',
} as const;

export interface PostListItem {
  id: string;
  type: PostListItemType;
  title: string;
  author: Author;
  score: number;
  commentCount: number;
  createdAt: string;
  userVoted?: boolean;
  /** @nullable */
  url?: string | null;
}

export interface FeedResponse {
  items: PostListItem[];
  page: number;
  pageSize: number;
  hasMore: boolean;
}

export type PostInputType = typeof PostInputType[keyof typeof PostInputType];


export const PostInputType = {
  discuss: 'discuss',
  ask: 'ask',
  show: 'show',
} as const;

export interface PostInput {
  type: PostInputType;
  /**
     * @minLength 3
     * @maxLength 300
     */
  title: string;
  /** @maxLength 50000 */
  content: string;
  /**
     * @maxLength 2000
     * @nullable
     */
  url?: string | null;
  /** @nullable */
  mediaPath?: string | null;
}

export interface PostUpdate {
  /**
     * @minLength 3
     * @maxLength 300
     */
  title?: string;
  /** @maxLength 50000 */
  content?: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  mediaPath?: string | null;
}

export interface Media {
  path: string;
  contentType: string;
  altText: string;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
}

export type Post = PostListItem & ({
  content: string;
  updatedAt: string;
  media?: Media | null;
});

export interface Comment {
  id: string;
  content: string;
  author: Author;
  score: number;
  createdAt: string;
  updatedAt: string;
  deleted: boolean;
  depth: number;
  userVoted: boolean;
}

export type PostDetail = Post & {
  comments: Comment[];
};

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 20000
     */
  content: string;
  /** @nullable */
  parentId?: string | null;
}

export interface CommentUpdate {
  /**
     * @minLength 1
     * @maxLength 20000
     */
  content: string;
}

export interface VoteResponse {
  voted: boolean;
  score: number;
}

export interface Profile {
  user: User;
  postCount: number;
  commentCount: number;
  karma: number;
  posts: PostListItem[];
  comments: Comment[];
}

export type ReportInputReason = typeof ReportInputReason[keyof typeof ReportInputReason];


export const ReportInputReason = {
  spam: 'spam',
  harassment: 'harassment',
  illegal: 'illegal',
  nsfw: 'nsfw',
  other: 'other',
} as const;

export type ReportInputTargetType = typeof ReportInputTargetType[keyof typeof ReportInputTargetType];


export const ReportInputTargetType = {
  post: 'post',
  comment: 'comment',
} as const;

export interface ReportInput {
  reason: ReportInputReason;
  targetType: ReportInputTargetType;
  targetId: string;
  /**
     * @maxLength 1000
     * @nullable
     */
  details?: string | null;
}

export type UploadInputContentType = typeof UploadInputContentType[keyof typeof UploadInputContentType];


export const UploadInputContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
  'image/gif': 'image/gif',
} as const;

export interface UploadInput {
  /** @maxLength 255 */
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: UploadInputContentType;
}

export interface UploadResponse {
  uploadUrl: string;
  objectPath: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type SectionParameter = typeof SectionParameter[keyof typeof SectionParameter];


export const SectionParameter = {
  top: 'top',
  new: 'new',
  ask: 'ask',
  show: 'show',
} as const;

export type PageParameter = number;

export type PageSizeParameter = number;

export type GetFeedParams = {
section?: SectionParameter;
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 50
 */
pageSize?: PageSizeParameter;
};

export type SearchParams = {
/**
 * @minLength 2
 * @maxLength 100
 */
q: string;
};