# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned media upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Presigned upload information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
components:
...
  schemas:
...
    UploadInput:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          maxLength: 255
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp, image/gif]
    UploadResponse:
      type: object
      required: [uploadUrl, objectPath]
      properties:
        uploadUrl:
          type: string
          format: uri
        objectPath:
          type: string